error handling

+++ app/app.js
@@ -35,7 +35,11 @@ app.get(`/api/${apiVer}/users/:id`, (req, res) => {
     const db = new sqlite3.Database(dbPath);
     const id = req.params.id;
     db.get(`SELECT * FROM users WHERE id = ${id}`, (err, row) => {
-        res.json(row);
+        if (!row) {
+            res.status(404).send({error: "Not Found!"});
+        } else {
+            res.status(200).json(row);
+        }
     });
     db.close();
 });
@@ -50,14 +54,12 @@ app.get(`/api/${apiVer}/search`, (req, res) => {
     db.close();
 });
 
-const run = async(sql, db, res, message) => {
+const run = async(sql, db) => {
     return new Promise((resolve, rejects) => {
         db.run(sql, (err) => {
             if (err) {
-                res.status(500).send(err)
-                return rejects();
+                return rejects(err);
             } else {
-                res.json({message: message});
                 return resolve();
             }
         });
@@ -66,42 +68,61 @@ const run = async(sql, db, res, message) => {
 
 // create new user
 app.post(`/api/${apiVer}/users`, async(req, res) => {
-    const db = new sqlite3.Database(dbPath);
-
-    const name = req.body.name;
-    const profile = req.body.profile ? req.body.profile : "";
-    const dateOfBirth = req.body.date_of_birth ? req.body.date_of_birth : "";
-
-    await run(
-        `INSERT INTO users (name, profile, date_of_birth) VALUES ("${name}", "${profile}", "${dateOfBirth}")`,
-        db, 
-        res, 
-        "新規ユーザを作成しました。"
-    );
-
-    db.close();
-});
-
-// update user data
-app.put(`/api/${apiVer}/users/:id`, async(req, res) => {
-    const db = new sqlite3.Database(dbPath);
-    const id = req.params.id;
-
-    // 現在のユーザ情報を取得
-    db.get(`SELECT * FROM users WHERE id = ${id}`, async(err, row) => {
-        const name = req.body.name ? req.body.name : row.name;
-        const profile = req.body.profile ? req.body.profile : row.profile;
-        const dateOfBirth = req.body.date_of_birth ? req.body.date_of_birth : row.date_of_birth;
-
-        await run(
-            `UPDATE users SET name="${name}", profile="${profile}", date_of_birth="${dateOfBirth}" WHERE id=${id}`,
-            db,
-            res,
-            "ユーザ情報を更新しました。",
-        )
-    });
+    if (!req.body.name || req.body.name === "") {
+        res.status(400).send({error: "ユーザ名が指定されていません。"});
+    } else {
+        const db = new sqlite3.Database(dbPath);
+        
+        const name = req.body.name;
+        const profile = req.body.profile ? req.body.profile : "";
+        const dateOfBirth = req.body.date_of_birth ? req.body.date_of_birth : "";
     
-    db.close();
+        try {
+            await run(
+            `INSERT INTO users (name, profile, date_of_birth) VALUES ("${name}", "${profile}", "${dateOfBirth}")`,
+            db,
+            );
+            res.status(201).send({ message: "新規ユーザを作成しました。"});
+        } catch (e) {
+            res.status(500).send({error: e});
+        }
+        
+        db.close();
+    }
+});
+    
+    // update user data
+app.put(`/api/${apiVer}/users/:id`, async(req, res) => {
+    if (!req.body.name || req.body.name === "") {
+        res.status(400).send({error: "ユーザ名が指定されていません。"});
+    } else {
+        const db = new sqlite3.Database(dbPath);
+        const id = req.params.id;
+        // 現在のユーザ情報を取得
+        db.get(`SELECT * FROM users WHERE id = ${id}`, async(err, row) => {
+            if (!row) {
+                res.status(404).send({ error: "指定したユーザが見つかりません。"});
+            } else {
+                const name = req.body.name ? req.body.name : row.name;
+                const profile = req.body.profile ? req.body.profile : row.profile;
+                const dateOfBirth = req.body.date_of_birth ? req.body.date_of_birth : row.date_of_birth;
+    
+                try {
+                    await run(
+                        `UPDATE users SET name="${name}", profile="${profile}", date_of_birth="${dateOfBirth}" WHERE id=${id}`,
+                        db,
+                        res,
+                        "ユーザ情報を更新しました。",
+                    );
+                    res.status(200).send({ message: "ユーザ情報を更新しました。"});
+                } catch (e) {
+                    res.status(500).send({ error: e});
+                }
+
+            }
+        });
+        db.close();
+    }
 });
 
 // delete user data
@@ -109,13 +130,22 @@ app.delete(`/api/${apiVer}/users/:id`, async(req, res) => {
     const db = new sqlite3.Database(dbPath);
     const id = req.params.id;
 
-    await run(
-        `DELETE FROM users WHERE id=${id}`,
-        db,
-        res,
-        "ユーザ情報を削除しました。",
-    )
-    
+    db.get(`SELECT * FROM users WHERE id = ${id}`, async(err, row) => {
+        if (!row) {
+            res.status(404).send({ error: "指定したユーザが見つかりません。"});
+        } else {
+            try {
+                await run(
+                    `DELETE FROM users WHERE id=${id}`,
+                    db
+                )
+                res.status(200).send({ message: "ユーザを削除しました。"});
+            } catch (e) {
+                res.status(500).send({ error: e});
+            }
+            
+        }
+    });
     db.close();
 });
